:recycle: Migrate to monorepo structure with Turborepo + restructure package
It seemed like the src code was lacking boundaries, so it felt appropriate to split the code into relevant code folders to keep the code siloed. There could likely be some additional polish made off of this migration, but this seems to be a large step in the direction for a codebase that is easier for contributors to add to.